Added some improvements.
Now the load balancer in use can be easily changed by modifying the generic parameter of the "run" method of the Server thanks to the introduction of Generics.

The json configuration file has been added to introduce the socket addresses (and relative weight) of the servers that will be used by the load balancer.